% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{cellTypeGOCorr}
\alias{cellTypeGOCorr}
\title{calculate correlation between cell types represented by scaled GO, per-species}
\usage{
cellTypeGOCorr(cell_type_go, corr_method = "pearson")
}
\arguments{
\item{cell_type_go}{cell type GO table calculated via getCellTypeGO}

\item{corr_method}{correlation method, choose among "pearson", "kendall", "spearman", default 'pearson'}
}
\value{
a dataframe with correlation between cell types
}
\description{
calculate correlation between cell types represented by scaled GO, per-species
}
\examples{
\dontrun{
cellTypeGOCorr(cell_type_go, corr_method = "pearson")
}
}
