% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{crossSpeciesCellTypeGOCorr}
\alias{crossSpeciesCellTypeGOCorr}
\title{calculate cross-species correlation between cell types represented by scaled GO}
\usage{
crossSpeciesCellTypeGOCorr(
  species_1,
  species_2,
  cell_type_go_sp1,
  cell_type_go_sp2,
  corr_method = "pearson"
)
}
\arguments{
\item{species_1}{name of species one}

\item{species_2}{name of species two}

\item{cell_type_go_sp1}{cell type GO table of species one calculated via getCellTypeGO}

\item{cell_type_go_sp2}{cell type GO table of species two calculated via getCellTypeGO}

\item{corr_method}{correlation method, choose among "pearson", "kendall", "spearman", default 'pearson'}
}
\value{
correlation between cell types
}
\description{
calculate cross-species correlation between cell types represented by scaled GO
}
\examples{
\dontrun{
crossSpeciesCellTypeGOCorr(species_1, species_2, go_sp1, go_sp2)
}
}
