% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dme_subset}
\alias{dme_subset}
\title{Drosophila gut scRNA-seq data, 10X Chromium
Subset to 45 cells per cell type as an example data}
\format{
a `Seurat` object
}
\source{
<https://flycellatlas.org/>
}
\usage{
dme_subset
}
\description{
Drosophila gut scRNA-seq data, 10X Chromium
Subset to 45 cells per cell type as an example data
}
\keyword{datasets}
