% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dme_tbl}
\alias{dme_tbl}
\title{Drosophila EMSEMBL to GO BP terms mapping table
Drosophila melanogaster}
\format{
a data frame
}
\source{
<https://www.ensembl.org/index.html>
}
\usage{
dme_tbl
}
\description{
Drosophila EMSEMBL to GO BP terms mapping table
Drosophila melanogaster
}
\keyword{datasets}
