% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{getCellTypeSharedTerms}
\alias{getCellTypeSharedTerms}
\title{query co-up and co-down regulated GO terms from certain cell type pairs}
\usage{
getCellTypeSharedTerms(
  shared_go,
  cell_type_sp1,
  cell_type_sp2,
  return_full = FALSE
)
}
\arguments{
\item{shared_go}{cell type shared GO table from getCellTypeSharedGO}

\item{cell_type_sp1}{cell type from sp1 to query}

\item{cell_type_sp2}{cell type from sp2 to query}

\item{return_full}{if return also pvals and logfc info, default FALSE}
}
\value{
a dataframe displaying co-up or co-down regulated GO terms for the queried cell type pair
}
\description{
query co-up and co-down regulated GO terms from certain cell type pairs
}
\examples{
\dontrun{
getCellTypeSharedTerms(shared_go, cell_type_sp1, cell_type_sp2)
}
}
