% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{globalvariables}
\alias{globalvariables}
\title{record some global variables: pre-defined column name in biomaRt query and markers}
\description{
record some global variables: pre-defined column name in biomaRt query and markers
}
