% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu_subset}
\alias{mmu_subset}
\title{Mouse stomach and intestine scRNA-seq data, microwell-seq
Subset to 50 cells per cell type as an example data}
\format{
a `Seurat` object
}
\source{
<https://bis.zju.edu.cn/MCA/>
}
\usage{
mmu_subset
}
\description{
Mouse stomach and intestine scRNA-seq data, microwell-seq
Subset to 50 cells per cell type as an example data
}
\keyword{datasets}
