% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu_tbl}
\alias{mmu_tbl}
\title{Mouse EMSEMBL to GO BP terms mapping table
Mus Musculus}
\format{
a data frame
}
\source{
<https://www.ensembl.org/index.html>
}
\usage{
mmu_tbl
}
\description{
Mouse EMSEMBL to GO BP terms mapping table
Mus Musculus
}
\keyword{datasets}
