% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{plotCellTypeCorrHeatmap}
\alias{plotCellTypeCorrHeatmap}
\title{plot clustered heatmap for cell type corr}
\usage{
plotCellTypeCorrHeatmap(corr_matrix, ...)
}
\arguments{
\item{corr_matrix}{correlation matrix from cellTypeGOCorr or crossSpeciesCellTypeGOCorr}

\item{...}{params to pass to slanter::sheatmap}
}
\value{
a sheatmap heatmap
}
\description{
plot clustered heatmap for cell type corr
}
\examples{
\dontrun{
plotCellTypeCorrHeatmap(corr_matrix, 'column')
}
}
