% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{plotCellTypeSankey}
\alias{plotCellTypeSankey}
\title{plot Sankey diagram for cell type links above a certain threshould}
\usage{
plotCellTypeSankey(corr_matrix, corr_threshould = 0.1, ...)
}
\arguments{
\item{corr_matrix}{cell type corr matrix from crossSpeciesCellTypeGOCorr}

\item{corr_threshould}{minimum corr value for positively related cell types, default 0.6}

\item{...}{additional params for sankeyNetwork}
}
\value{
a Sankey plot showing related cell types
}
\description{
plot Sankey diagram for cell type links above a certain threshould
}
\examples{
\dontrun{
plotCellTypeSankey(corr_matrix, 0.1)
}
}
