% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dme_tbl}
\alias{dme_tbl}
\title{Drosophila EMSEMBL gene and GO annotation, subset to genes present in `dme_subset`}
\format{
a `data.frame` object
}
\source{
<http://www.ensembl.org/>
}
\usage{
dme_tbl
}
\description{
Drosophila EMSEMBL gene and GO annotation, subset to genes present in `dme_subset`
}
\keyword{datasets}
