% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmu_tbl}
\alias{mmu_tbl}
\title{Mouse EMSEMBL gene and GO annotation, subset to genes present in `mmu_subset`}
\format{
a `data.frame` object
}
\source{
<http://www.ensembl.org/>
}
\usage{
mmu_tbl
}
\description{
Mouse EMSEMBL gene and GO annotation, subset to genes present in `mmu_subset`
}
\keyword{datasets}
